+++ frontend/src/auth/sign-up/ui/index.js
@@ -1,0 +1,17 @@
+import React, { useState } from "react";
+import UserForm from "../../ui/UserForm";
+
+export const SignUp = () => {
+  return (
+    <div>
+      <UserForm
+        firstName={userState.firstName}
+        lastName={userState.lastName}
+        email={userState.email}
+        password={userState.password}
+        onChange={changeHandler}
+        onSubmitUser={registerUser}
+      />
+    </div>
+  );
+};
