+++ frontend/src/user/ui/card/constants.js
@@ -1,0 +1,3 @@
+export const LOREM_TEXT = `
+Lorem Ipsum is simply dummy text of the printing and typesetting
+industry.`;
